refactor(web): simplify SomedayTab component and integrate useWeekLabel hook
